requirements.txt
anyio==4.6.0
asgiref==3.8.1
boto3==1.35.36
botocore==1.35.36
certifi==2024.8.30
charset-normalizer==3.4.0
defusedxml==0.7.1
Django==5.1.2
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.2
h11==0.14.0
httpcore==1.0.6
httpie==3.2.3
httpx==0.27.2
idna==3.10
jmespath==1.0.1
Markdown==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
ollama==0.3.3
pillow==10.4.0
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.9.2
s3transfer==0.10.3
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
urllib3==2.2.3
whitenoise==6.7.0
